Reorganizing the base implementation
- Hide the implementation details
- Remove a layer of indirection
