+++ src/baseStream.js
@@ -3,9 +3,7 @@ const EventEmitter = require('events');
 const { StreamError } = require('./errors');
 // const { strFromEvent } = require('./event');
 
-import type { Event } from './event';
-
-export type Push<T> = (Event<T>) => boolean;
+import type { Consumer, Event, Producer, Stream } from './basics';
 
 const PRODUCER_STATUS = {
   ONGOING: 'ONGOING',
@@ -13,7 +11,6 @@ const PRODUCER_STATUS = {
   ERROR: 'ERROR'
 };
 type ProducerStatus = $Keys<typeof PRODUCER_STATUS>;
-export type Producer<T> = (push: Push<T>, stop?: boolean) => Promise<void> | void;
 
 const CONSUMER_STATUS = {
   NONE: 'NONE',
@@ -22,18 +19,17 @@ const CONSUMER_STATUS = {
   DONE: 'DONE'
 };
 type ConsumerStatus = $Keys<typeof CONSUMER_STATUS>;
-export type Consumer<T> = (Event<T>) => Promise<boolean> | boolean;
 
-export type InternalStreamConfiguration = {|
+export type Configuration = {|
   bufferHighWaterMark: number
 |}
 
-const DEFAULT_INTERNAL_STREAM_CONFIGURATION : InternalStreamConfiguration = {
+const DEFAULT_INTERNAL_STREAM_CONFIGURATION : Configuration = {
   bufferHighWaterMark: 100
 };
 
-class InternalStream<T> extends EventEmitter {
-  cfg: InternalStreamConfiguration;
+class BaseStream<T> extends EventEmitter implements Stream<T> {
+  cfg: Configuration;
 
   producer: ?Producer<T>;
   producerStatus: ProducerStatus;
@@ -42,7 +38,7 @@ class InternalStream<T> extends EventEmitter {
   consumer: ?Consumer<T>;
   consumerStatus: ConsumerStatus;
 
-  constructor(producer: ?Producer<T>, cfg: InternalStreamConfiguration = DEFAULT_INTERNAL_STREAM_CONFIGURATION) {
+  constructor(producer: ?Producer<T>, cfg: Configuration = DEFAULT_INTERNAL_STREAM_CONFIGURATION) {
     super();
     this.producerStatus = PRODUCER_STATUS.ONGOING;
     this.producer = producer;
@@ -79,7 +75,7 @@ class InternalStream<T> extends EventEmitter {
     }
   }
   _consume(event: Event<T>) {
-    //console.log('InternalStream._consume', strFromEvent(event), this.consumerStatus);
+    //console.log('BaseStream._consume', strFromEvent(event), this.consumerStatus);
     // By construction we're sure that
     //  - `this.consumerStatus == CONSUMER_STATUS.READY`
     // if (this.consumerStatus != CONSUMER_STATUS.READY) {
@@ -109,7 +105,7 @@ class InternalStream<T> extends EventEmitter {
     }
   }
   _produce(event: Event<T>): boolean {
-    //console.log('InternalStream._produce', strFromEvent(event));
+    //console.log('BaseStream._produce', strFromEvent(event));
     let productionDone = false;
     if (this.producerStatus != PRODUCER_STATUS.ONGOING) {
       throw new StreamError('No event should be produced once the stream has ended.');
@@ -144,7 +140,7 @@ class InternalStream<T> extends EventEmitter {
     }
   }
   _doConsume() {
-    //console.log('InternalStream._doConsume', this.consumerStatus, this.buffer.length);
+    //console.log('BaseStream._doConsume', this.consumerStatus, this.buffer.length);
 
     // 1 - let's evacuate what is in the buffer
     while (
@@ -178,7 +174,7 @@ class InternalStream<T> extends EventEmitter {
     }
   }
   consume(consumer: Consumer<T>): Promise<void> {
-    // console.log('InternalStream.consume');
+    // console.log('BaseStream.consume');
     if (this.consumer != null) {
       throw new StreamError('Stream already being consumed.');
     }
@@ -196,18 +192,25 @@ class InternalStream<T> extends EventEmitter {
       this._doConsume();
     });
   }
-  waitAndPush(event: Event<T>): Promise<boolean> | boolean {
-    // console.log('InternalStream.waitAndPush', strFromEvent(event), this.consumerStatus);
+  push(event: Event<T>): Promise<boolean> | boolean {
+    // console.log('BaseStream.push', strFromEvent(event), this.consumerStatus);
     switch (this.consumerStatus) {
       case CONSUMER_STATUS.BUSY:
         return new Promise(this.once.bind(this, 'consumerReady'))
-          .then(this.waitAndPush.bind(this, event));
+          .then(this.push.bind(this, event));
       default:
         return this._produce(event);
     }
   }
+  thru<R, Fn: (BaseStream<T>) => R>(f: Fn): R {
+    return f(this);
+  }
+}
+
+function createBaseStream<T>(producer: ?Producer<T>, cfg: Configuration = DEFAULT_INTERNAL_STREAM_CONFIGURATION): Stream<T> {
+  return new BaseStream(producer, cfg);
 }
 
 module.exports = {
-  InternalStream
+  createBaseStream
 };
