+++ src/index.js
@@ -1,6 +1,8 @@
 // @flow
 const { drain }  = require('./drain');
-const { from, of, produce, throwError, transduce } = require('./basics');
+const { produce } = require('./produce');
+const { transduce } = require('./transduce');
+const { from, of, throwError } = require('./from');
 const { map }  = require('./map');
 const { reduce }  = require('./reduce');
 const { subscribe }  = require('./subscribe');
